Commit 4

+++ Capstone/dotnet/Capstone/Candy.cs
@@ -6,7 +6,7 @@ namespace Capstone
 {
     public class Candy : Item
     {
-        public Candy(string identifier, string name, double price, string type, int stock) : base(identifier, name, price, type, stock)
+        public Candy(string identifier, string name, decimal price, string type, int stock) : base(identifier, name, price, type, stock)
         {
 
         }
