+++ Capstone/dotnet/Capstone/Gum.cs
@@ -6,7 +6,7 @@ namespace Capstone
 {
     class Gum : Item
     {
-        public Gum(string identifier, string name, double price, string type, int stock) : base(identifier, name, price, type, stock)
+        public Gum(string identifier, string name, decimal price, string type, int stock) : base(identifier, name, price, type, stock)
         {
 
         }
